extracted nats connection builder on other file
